Add tests for new apply page
